rearrange everything to get rspec test to run
